진행률 기록 다이어로그 ui
